+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feedback bij een NPC toegevoegd over de werking
 - Zeer veel nieuwe houserobbery locaties toegevoegd
 - Locate player/vehicle commands toegevoegd aan adminmenu
+- Actieve melding om weer te geven dat je in cloak/noclip ben
 
 ### Changed
 - Bait price aangepast